Break up Generator into modules
